Create cnn_multicombs.py
Script for training CNNs on imagined speech EEG data which has been 
divided into two classes, words and vowels. Data is loaded, separated into the 
two classes and labels of 0 and 1 assigned. 

This means that a model is trained on 'arriba' vs '/a/', then 'arriba' vs /e/, ..., 'Izquierda' vs '/u/'